backend-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "backend-service"
version = "0.1.0"
description = "Backend service for similar-screens solution"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "contracts",
    "minio",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "qdrant-client",
    "httpx",
    "pillow",
]

[tool.setuptools]
packages = ["backend_service"]
